+++ docs/changelog.md
@@ -4,6 +4,10 @@
 The following sections describe major changes per version 
 and can be helpful with version upgrades.
 
+## 0.2 
+
+- Adds a basic NodeDriver test helper and JUnit5 extension.
+
 ## 0.1 
 
-- Initial release, adds Cordform containers and JUnit5 extention
+- Initial release, adds Cordform-based testcontainers and JUnit5 extension.